django_app/useractivity/requirements.txt
asgiref==3.2.3
attrs==19.3.0
Automat==20.2.0
bs4==0.0.1
bson==0.5.10
certifi==2019.11.28
chardet==3.0.4
constantly==15.1.0
Django==3.0.2
django-compat==1.0.15
django-crispy-forms==1.9.0
djangorestframework==3.11.1
dnspython==1.16.0
html-parser==0.2
HTMLParser==0.0.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
inscriptis==1.0
logger==1.4
lxml==4.5.0
mailbox==0.4
ply==3.11
py3dns==3.2.1
PyHamcrest==2.0.2
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-toolbelt==0.9.1
six==1.14.0
socks==0
soupsieve==1.9.5
SQLAlchemy==1.3.13
sqlparse==0.3.0
urllib3==1.25.8
validate-email==1.3
zope.interface==4.7.1
